Projeto Finalizado

+++ src/main/java/com/gerenciador/tarefas/service/GerenciadorTarefasService.java
@@ -1,6 +1,9 @@
 package com.gerenciador.tarefas.service;
 
 import com.gerenciador.tarefas.entity.Tarefa;
+import com.gerenciador.tarefas.excecoes.NaoPermitirExcluirException;
+import com.gerenciador.tarefas.excecoes.NaoPermitoAlterarStatusException;
+import com.gerenciador.tarefas.excecoes.TarefaExistenteException;
 import com.gerenciador.tarefas.repository.GerenciadorTarefasRepository;
 import com.gerenciador.tarefas.request.AtualizarTarefaRequest;
 import com.gerenciador.tarefas.request.CadastrarTarefaRequest;
@@ -25,6 +28,12 @@ public class GerenciadorTarefasService {
 
     public Tarefa salvarTarefa(CadastrarTarefaRequest request) {
 
+        Tarefa tarefaValidacao = gerenciadorTarefasRepository.findByTituloOrDescricao(request.getTitulo(), request.getDescricao());
+
+        if (tarefaValidacao != null) {
+            throw new TarefaExistenteException("Já existe uma tarefa com o mesmo titulo ou descrição");
+        }
+
         Tarefa tarefa = Tarefa.builder()
                 .quantidadeHorasEstimadas(request.getQuantidadeHorasEstimadas())
                 .status(TarefaStatusEnum.CRIADA)
@@ -38,17 +47,29 @@ public class GerenciadorTarefasService {
     }
 
     public Page<Tarefa> obtemTarefasPorTitulo(String titulo, Pageable pegeable){
-       return this.gerenciadorTarefasRepository.findByTituloContaining(titulo, pegeable);
+       return this.gerenciadorTarefasRepository.findByTituloContainingOrderByDataAtualizacaoDesc(titulo, pegeable);
     }
 
     public Page<Tarefa> obtemTodasTarefas(Pageable pegeable){
 
-        return this.gerenciadorTarefasRepository.findAll(pegeable);
+        return this.gerenciadorTarefasRepository.findAllByOrderByDataAtualizacaoDesc(pegeable);
     }
 
     public Tarefa atualizarTarefa(Long id, AtualizarTarefaRequest request){
 
         Tarefa tarefa = this.gerenciadorTarefasRepository.findById(id).get();
+
+        if (tarefa.getStatus().equals(TarefaStatusEnum.FINALIZADA)) {
+            throw new NaoPermitoAlterarStatusException("Não permitido mover tarefa que está FINALIZADA");
+        }
+
+        if (tarefa.getStatus().equals(TarefaStatusEnum.CRIADA) && request.getStatus().equals(TarefaStatusEnum.FINALIZADA)) {
+          throw new NaoPermitoAlterarStatusException("Não permitido mover a tarefa para FINALIZADA se a mesma estiver com status de CRIADA");
+        }
+
+        if (tarefa.getStatus().equals(TarefaStatusEnum.BLOQUEADA) && request.getStatus().equals(TarefaStatusEnum.FINALIZADA)) {
+           throw new NaoPermitoAlterarStatusException("Não permitido mover a tarefa para FINALIZADA se a mesma estiver com status de BLOQUEADA");
+        }
 
         tarefa.setQuantidadeHorasEstimadas(request.getQuantidadeHorasEstimadas());
         tarefa.setStatus(request.getStatus());
@@ -63,6 +84,13 @@ public class GerenciadorTarefasService {
     }
 
     public void excluirTarefa(Long id){
+
+        Tarefa tarefa = this.gerenciadorTarefasRepository.findById(id).get();
+
+        if (!TarefaStatusEnum.CRIADA.equals(tarefa.getStatus())) {
+            throw new NaoPermitirExcluirException();
+        }
+
         this.gerenciadorTarefasRepository.deleteById(id);
     }
 }
